+++ src/main/kotlin/com/github/intellij/plugins/mt4ij/config/SettingsConfigurable.kt
@@ -3,12 +3,12 @@ package com.github.intellij.plugins.mt4ij.config
 import com.github.intellij.plugins.mt4ij.ApiHelpers
 import com.github.intellij.plugins.mt4ij.Bundle
 import com.github.intellij.plugins.mt4ij.activities.ProjectScannerActivity
-import com.intellij.openapi.options.Configurable
+import com.intellij.openapi.options.SearchableConfigurable
 import com.intellij.openapi.project.Project
 import javax.swing.JComponent
 
 /*
-    // Swing
+    // Swing UI
 
     https://stackoverflow.com/a/29258675/418599
 
@@ -16,11 +16,13 @@ import javax.swing.JComponent
 
     https://intellij-support.jetbrains.com/hc/en-us/community/posts/360003316280-Adding-Your-Plugin-Setting-to-the-IDE-Settings-Dialog
     https://programtalk.com/vs/intellij-haxe/src/common/com/intellij/plugins/haxe/config/HaxeSettingsConfigurable.java/
+
+    // Searchable label
+
+    https://intellij-support.jetbrains.com/hc/en-us/community/posts/206133719-How-does-SearchableConfigurable-work
  */
 
- //TODO Make label searchable
-
-internal class SettingsConfigurable(project: Project) : Configurable {
+internal class SettingsConfigurable(project: Project) : SearchableConfigurable {
     private val projectRef   = project
     private val settingsForm = SettingsForm()
 
@@ -53,5 +55,9 @@ internal class SettingsConfigurable(project: Project) : Configurable {
 
     override fun getDisplayName(): String {
         return Bundle.message("mt4ij.settings.name")
+    }
+
+    override fun getId(): String {
+        return displayName
     }
 }